requirements.txt
streamlit
pyaudio
wave
openai
google.generativeai
python-dotenv
